Adición: Movimiento de código de noción espacial a otro archivo
Se creó un archivo adicional para el código relacionado a la noción espacial.
